Create security App for women
A passenger can register as user. only they get registered all their details will be shared with the railway trusted contact. In the case the passenger is facing any problem she can lodge complaint regarding the problem in the app.so the problem is notices by the authority and the trusted contacts. Then they can take immediate action. if needed they can take assistance of the police. once the passenger reaches to their destination immediately both railway authority and trusted contacts will be informed that the passenger is safe.